requirements.txt
transformers
datasets
accelerate
sentencepiece
sentence-transformers
scikit-learn
evaluate
huggingface_hub
streamlit
python-dotenv
pandas
numpy
matplotlib
